# Reduce generated IL code size