--- a/src/XamlX/IL/Emitters/MarkupExtensionEmitter.cs
+++ b/src/XamlX/IL/Emitters/MarkupExtensionEmitter.cs
@@ -36,7 +36,7 @@ namespace XamlX.IL.Emitters
             
             if (me.ProvideValue.Parameters.Count > 0)
                 ilgen
-                    .Emit(OpCodes.Ldloc, context.ContextLocal);
+                    .Ldloc(context.ContextLocal);
 
             if (needProvideValueTarget)
             {