--- a/src/XamlX/IL/IXamlILLocal.cs
+++ b/src/XamlX/IL/IXamlILLocal.cs
@@ -1,0 +1,12 @@
+﻿using XamlX.TypeSystem;
+
+namespace XamlX.IL
+{
+#if !XAMLX_INTERNAL
+    public
+#endif
+    interface IXamlILLocal : IXamlLocal
+    {
+        int Index { get; }
+    }
+}