… src/XamlX/IL/SreTypeSystem.cs
… src/XamlX/IL/XamlILEmitterExtensions.cs
… src/XamlX/IL/XamlIlCompiler.cs
… src/XamlX/Transform/TransformerConfiguration.cs
… src/XamlX/Transform/Transformers/ResolvePropertyValueAddersTransformer.cs
… src/XamlX/Transform/XamlTransformHelpers.cs
… src/XamlX/TypeSystem/TypeSystem.cs
… tests/XamlParserTests/DynamicSettersTests.cs
… tests/XamlParserTests/IntrinsicsTests.cs
… tests/XamlParserTests/ServiceProviderTests.cs