I was poking around the IL generated from XAML files by Avalonia to understand how the IL compiler works, and noticed some redundant instructions. So here's a pull request trying to optimize those a bit :) The goal is to reduce the final IL size, and if possible improve the JIT compiling times, [which are high](https://github.com/AvaloniaUI/Avalonia/issues/5666#issuecomment-799408016).

## Notes

`Avalonia.Themes.Fluent` was the main target of these optimizations since it has many, relatively heavy, XAML files.

This PR contains several independent commits and I think it's easier to review commit per commit. If needed, I can split it into several PRs. 

`LangVersion` is set to `latest` to use `switch` expressions and improved pattern matching. If that's not acceptable, I can rewrite those with older C# constructs.

Hopefully no breaking changes in public, non-IL specific types (`IXamlPropertySetter`, etc.). However there are new overloads in `IXamlILEmitter`, which is technically breaking, but is very unlikely to be implemented by types other than those already in XamlX.

## Commits

### Fixed x:Null incorrectly being assignable to a value type for Cecil

Not an optimization, but a fix for a bug I stumbled upon later when optimizing the property setters. For Cecil, `x:Null` was always incorrectly assignable to everything, even value types. It now has the same logic as S.R.Emit. A failing test has been added, which now passes.

### Using optimized ldloc/stloc/ldarg/ldc.i4 opcodes when possible

A low hanging fruit: use the short encoding forms of `ldloc`/`stloc`/`ldc.i4` when possible. This reduces most `ldloc`/`stloc` from 5 bytes to 1. SRE already does this automatically internally, but Cecil doesn't.

### Removed unnecessary casts for dynamic property assignments

`PropertyAssignmentEmitter` was adding an unneeded cast right after a value was checked to either be statically assignable by downcast, or after a runtime `isinst` call, which pushes the correct the type on the stack on success.

### Removed unnecessary casts for adders

Casts were present on base interface calls for adders, which was very noticeable for Avalonia's `Style.Resources` which is typed as `IResourceDictionary`, but whose `Add` method is on `IDictionary<TKey,TValue>`. This removes those as they're not necessary. (Even `IAddChild` logic, which is right after, didn't add those casts.)

### NamespaceInfoProvider: generated CreateNamespaceInfo method to avoid duplicate code

This commits generates a simple helper function to reduce highly duplicated code when generating the namespaces alias mappings.

### Cached dynamic property setters

The biggest change in this PR. When `PropertyAssignmentEmitter` has to do runtime dispatch because there are multiple setters that can't be resolved at compile time, it now generates a single reusable method for those setters, per XAML file. For `Avalonia.Themes.Fluent`, all `<SolidColorBrush x:Key="xxx" Color="{StaticResource yyy}" />` in a single file calls an unique method instead of generating the same branches that check if the static resource returned an `UnsetValueType`, `IBinding` or `Color`. The result is 8 different methods in the assembly instead of 183 times the branches.

This commit requires that `IXamlPropertySetter` correctly implements `Equals` and `GetHashCode`. If an implementation doesn't, the generated code is still correc: the generated method simply won't be reused (so no IL size gain). Ideally `IXamlPropertySetter` should inherit from `IEquatable<IXamlPropertySetter>` but that would be a breaking change. If this PR is accepted, I have a follow up one on Avalonia for its custom setter implementations.

`GetHashCode` is implemented on `PropertySetterBinderParameters` despite this class being mutable, but it doesn't matter in practice: binding parameters are never changed once assigned in existing code (and doing so would probably break things).

While writing this, I noticed that `AllowRuntimeNull` wasn't really working. Each generated branch is checking for the value to be of a specific type, so `null` won't ever match even when this property is `true`. This is fixed.

### Property setters can emit their own arguments for better IL generation

`IXamlPropertySetter.Emit` is called with arguments already being on the stack. Most custom setters need to push extra arguments before, and thus resort to `stloc` all existing arguments, push what they need, then `ldloc` the arguments. This is really noticeable on adders. A new `IXamlILOptimizedEmitablePropertySetter` interface has been added that setters can optionally implement, which must emit the arguments itself.

If this PR is accepted, I have a follow up one on Avalonia that implements this interface on the custom setters.

### Lazy NamespaceInfoProvider.XmlNamespaces

This commit doesn't decrease IL size, but avoids the work of creating the XML namespace dictionary on initialization. The `XmlNamespaces` property has a lazy implementation instead, which will be jitted only when needed. 

This also helps to reduce runtime memory allocations a little bit: in Avalonia those namespaces are only used in reflection bindings using either `FindAncestor` or having an attached property in their path.

The lazy initialization is lock free: it's very likely to be used on a single thread, if it isn't the race isn't problematic since the dictionary creation has no side effect, and the returned dictionary is read-only and never mutated.

Additionally, the returned dictionary uses an initial capacity and fixed size arrays, to avoid allocating unused memory.

## Results

### JIT

With `Avalonia.Themes.Fluent` stripped of its fonts (to keep only code), here are the results of this PR on my PC:

|               | Before   | After   | Diff   |
| :------------ | -------: | ------: | -----: |
| File Size     |   833 KB |  574 KB | -31.1% |
| Run (JIT)     |   293 ms |  254 ms | -13.3% |

The *Run (JIT)* column is the time spent in `AvaloniaLoader.Load(this)` in an empty App that has only `<FluentTheme />` in its styles. Average of 10 runs in Release mode, with the .NET 6.0.7 runtime.

Overall, some noticeable improvements in both file size and JIT time spent.

### R2R

Being curious, I also tested ReadyToRun (win-x64):

|               | Before   | After   | Diff   |
| :------------ | -------: | ------: | -----: |
| File Size R2R |  2501 KB | 1939 KB | -22.5% |
| Run (R2R)     |   186 ms |  186 ms |   0.0% |

As expected the PR doesn't improve the run time since the JIT is barely involved, but the final assembly still has a nice size reduction.