@kekekeks 2022-08-03
>A low hanging fruit: use the short encoding forms of ldloc/stloc/ldc.i4 when possible. This reduces most ldloc/stloc from 5 bytes to 1. SRE already does this automatically internally, but Cecil doesn't.

I think Cecil has a separate [Optimize method](https://github.com/jbevain/cecil/blob/master/rocks/Mono.Cecil.Rocks/MethodBodyRocks.cs#L177) for doing that automatically, but we aren't using it for some reason