@kekekeks 2022-08-03
Note, that due to the amount of changes I could only merge the PR after a corresponding PR in Avalonia repo passes the unit tests