@MrJul 2022-08-03
> > A low hanging fruit: use the short encoding forms of ldloc/stloc/ldc.i4 when possible. This reduces most ldloc/stloc from 5 bytes to 1. SRE already does this automatically internally, but Cecil doesn't.
> 
> I think Cecil has a separate [Optimize method](https://github.com/jbevain/cecil/blob/master/rocks/Mono.Cecil.Rocks/MethodBodyRocks.cs#L177) for doing that automatically, but we aren't using it for some reason

I wasn't aware about that method. Should it be used instead and the first commit removed?


> I don't have a lot to add as a review of this PR as it's not my area, but I would like to give you a 👏 for the best PR description I think I've ever seen from a new contributor ;)

Thank you! :)

> Note, that due to the amount of changes I could only merge the PR after a corresponding PR in Avalonia repo passes the unit tests

That's completely understandable, I'll add a PR soon. Will the CI be able to update the submodule since it will refer to a non-merged PR commit?