@kekekeks 2022-08-03
>Should it be used instead and the first commit removed?

We definitely should use it since it optimizes branching ops to short versions too.

>Will the CI be able to update the submodule since it will refer to a non-merged PR commit?

Commits in github forks are globally visible, so CI can pick submodule commit from any fork, yes