@MrJul 2022-08-04
I just pushed a new commit after doing more testing.

The `EmitConvert` call in was actually still needed for one specific case: `T` to `Nullable<T>`. It now uses the current type on the stack (after `isinst`) instead of the original value type, so extra casts aren't emitted in all other cases.

I also added new unit tests that check the correctness of dynamic setters for value types, nullable value types and reference types.