@MrJul 2022-08-05
New numbers for this PR with the empty app now that the (awesome) lazy resources PR has been merged in Avalonia:

|               | Before   | After   | Diff   |
| :------------ | -------: | ------: | -----: |
| File Size     |   1032 KB |  815 KB | -21.0% |
| Run (JIT)     |   78 ms |  57 ms | -27.0% |

However, with the lazy resources, setters aren't reused since each resource has its own closure class and the cache works per class, as mentioned earlier. I'd like to keep this PR code as-is if that's fine with the maintainers, and improve the setter cache part in another PR.

As a side note, the fluent theme XAML files now compile to more than 1700 classes to the assembly, which might also need a change (even though I haven't measured the impact of having that many types versus one big type with many methods).