@grokys 2022-08-05
> I'd like to keep this PR code as-is if that's fine with the maintainers, and improve the setter cache part in another PR.

Sounds good to me - `Setters` are still used as part of styles (which won't be affected by lazy resources).

> As a side note, the fluent theme XAML files now compile to more than 1700 classes to the assembly, which might also need a change

Yep, I'm pretty sure there's a lot of optimization we can do. We should check how much overhead this requires.